+++ src/components/molecules/Timeline/Timelines.stories.tsx
@@ -1,26 +1,19 @@
-import React, { FC } from "react";
-import { Meta } from "@storybook/react";
+import React from "react";
+import { Meta, StoryObj } from "@storybook/react";
 
 // Helpers
-import { args, propCategory } from "../../../../stories/assets/storybook.globals";
+import { args, propCategory, storyObjBuilder } from "../../../../stories/assets/storybook.globals";
 
 // Components
 import TimelinesComponent, { ITimelinesProps } from "./Timelines";
-import TimelinePoint from "./TimelinePoint";
+import TimelinePoint, { ITimelineProps } from "./TimelinePoint";
 
-const meta: Meta<typeof TimelinesComponent> = {
-    title: "Molecules/Timeline",
+const meta: Meta = {
+    title: "Molecules/Timelines",
     component: TimelinesComponent,
-    subcomponents: { TimelinePoint },
-    argTypes: {
-        className: args({ control: "false", ...propCategory.appearance }),
-        direction: args({ control: "select", ...propCategory.appearance }),
-        position: args({ control: "select", ...propCategory.appearance })
-    },
-    args: {
-        direction: "vertical",
-        position: "after"
-    } as ITimelinesProps
+    subcomponents: {
+        Timeline: TimelinePoint
+    }
 };
 
 export default meta;
@@ -33,13 +26,54 @@ const inlineData = [
     { title: "Task E", description: "Description E", status: "success" }
 ] as const;
 
-const Template: FC<ITimelinesProps> = (props) => {
-    return (
-        <TimelinesComponent {...props}>
-            {inlineData.map((timeline) => {
-                return <TimelinePoint {...timeline} />;
-            })}
-        </TimelinesComponent>
-    );
+type Story = StoryObj<ITimelinesProps>;
+type StoryTimeline = StoryObj<ITimelineProps>;
+
+export const Timelines: Story = {
+    argTypes: {
+        className: args({ control: "false", ...propCategory.appearance }),
+        direction: args({ control: "select", ...propCategory.appearance }),
+        position: args({ control: "select", ...propCategory.appearance })
+    },
+    args: {
+        direction: "vertical",
+        position: "after"
+    },
+    render: (props) => {
+        return (
+            <TimelinesComponent {...props}>
+                {inlineData.map((timeline) => {
+                    return <TimelinePoint {...timeline} />;
+                })}
+            </TimelinesComponent>
+        );
+    }
 };
-export const Timelines = Template.bind({});
+
+export const Timeline: StoryTimeline = storyObjBuilder({
+    argTypes: {
+        direction: args({ control: "select", ...propCategory.appearance }),
+        title: args({ control: "text", ...propCategory.content }),
+        status: args({
+            control: "select",
+            ...propCategory.content,
+            options: ["default", "active", "success", "error", "pending"]
+        }),
+        description: args({ control: "text", ...propCategory.content })
+    },
+    args: {
+        title: "Task A",
+        description: "Description A",
+        status: "success",
+        direction: "vertical"
+    },
+    render: (props) => {
+        const { direction } = props;
+        return (
+            <TimelinesComponent direction={direction}>
+                <TimelinePoint {...props} />
+                <TimelinePoint {...props} />
+            </TimelinesComponent>
+        );
+    }
+});
